+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add Progressive Web App feature
 -   Add paging to calendar mobile page
 -   Add Ubuntu bold to react pdf component
+-   Add event modal on mobile calendar component
 
 ## Fix